requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
langchain-text-splitters
pypdf
bs4
arxiv
pymupdf
lxml
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
watchdog
langchain
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
